One Hour Game Jam #026 - Adorable
Not much of a game but experimented with animations - sprite animations
in this case - which has always been something I have had issues with.
Did some 2D graphics, and physics, including using a capsule cast
for determining if the player is touching the ground. Use A+D for left
and right controls. Press S to melt. Character inspired by Putty on the
Amiga.